+++ src/main/java/frc/robot/commands/climber/ClimberDown.java
@@ -1,0 +1,28 @@
+package frc.robot.commands.climber;
+import edu.wpi.first.wpilibj2.command.Command;
+import frc.robot.Constants;
+import frc.robot.subsystems.ClimberSubsystem;
+public class ClimberDown extends Command {
+    private ClimberSubsystem climber;
+    public ClimberDown(ClimberSubsystem climber) {
+        this.climber = climber;
+    }
+    @Override
+    public void initialize() {   
+    }
+    @Override
+    public void execute() {
+        climber.setMotorVoltage(-Constants.ClimberConstants.climberVoltage);
+    }
+    @Override
+    public void end(boolean interrupted) {
+        climber.setMotorVoltage(0);
+    }
+    @Override 
+    public boolean isFinished() {
+        return false;
+    }
+
+
+
+}
