+++ src/main/java/frc/robot/commands/indexAndShoot/Index.java
@@ -2,21 +2,20 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.indexAndShoot;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.ShooterSubsystem;
+import frc.robot.Constants;
+import frc.robot.subsystems.IndexerSubsystem;
 
-public class Shoot extends Command {
-  /** Creates a new Shoot. */
-  private final ShooterSubsystem shooter;
-  private final double speed;
+public class Index extends Command {
+  /** Creates a new RunIntake. */
+  public final IndexerSubsystem indexer;
 
-  public Shoot(ShooterSubsystem shooter, double speed) {
-    this.shooter = shooter;
-    this.speed = speed;
-
-    addRequirements(this.shooter);
+  public Index(IndexerSubsystem indexer) {
+    // Use addRequirements() here to declare subsystem dependencies.
+    this.indexer = indexer;
+    addRequirements(indexer);
   }
 
   // Called when the command is initially scheduled.
@@ -26,13 +25,18 @@ public class Shoot extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    shooter.setLeftMotorVoltage(speed);
+    if(indexer.getSensorStatus() == false) {
+    indexer.setIntakeVoltage(Constants.IntakeConstants.intakeSpeed);
+    indexer.setFeederVoltage(Constants.IntakeConstants.indexSpeed);
+    } else {
+      indexer.stop();
+    }
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    shooter.setLeftMotorVoltage(0);
+    indexer.stop();
   }
 
   // Returns true when the command should end.
