+++ src/main/java/frc/robot/commands/indexAndShoot/Shoot.java
@@ -2,20 +2,21 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.indexAndShoot;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.Constants;
-import frc.robot.subsystems.IndexerSubsystem;
+import frc.robot.subsystems.ShooterSubsystem;
 
-public class Index extends Command {
-  /** Creates a new RunIntake. */
-  public final IndexerSubsystem indexer;
+public class Shoot extends Command {
+  /** Creates a new Shoot. */
+  private final ShooterSubsystem shooter;
+  private final double speed;
 
-  public Index(IndexerSubsystem indexer) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.indexer = indexer;
-    addRequirements(indexer);
+  public Shoot(ShooterSubsystem shooter, double speed) {
+    this.shooter = shooter;
+    this.speed = speed;
+
+    addRequirements(this.shooter);
   }
 
   // Called when the command is initially scheduled.
@@ -25,18 +26,13 @@ public class Index extends Command {
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
-    if(indexer.getSensorStatus() == false) {
-    indexer.setIntakeVoltage(Constants.IntakeConstants.intakeSpeed);
-    indexer.setFeederVoltage(Constants.IntakeConstants.indexSpeed);
-    } else if (indexer.getSensorStatus() == true ) {
-      indexer.stop();
-    }
+    shooter.setLeftMotorVoltage(speed);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
-    indexer.stop();
+    shooter.setLeftMotorVoltage(0);
   }
 
   // Returns true when the command should end.
